Fixed file for user tab view and add new file U5_Congrats.swift
